chore: refactor name of file to match name of class
